requirements.txt
Django==5.2.4
djangorestframework==3.14.0
psycopg2-binary==2.9.9
openai==1.3.7
python-dotenv==1.0.0 